+++ src/my_first_adventure_game/game/scenes/__init__.py
@@ -1,4 +1,5 @@
+from my_first_adventure_game.game.scenes.defeat_scene import DefeatScene
 from my_first_adventure_game.game.scenes.gameplay_scene import GameplayScene
 from my_first_adventure_game.game.scenes.title_scene import TitleScene
 
-__all__ = ["GameplayScene", "TitleScene"]
+__all__ = ["DefeatScene", "GameplayScene", "TitleScene"]